chore: checkout code first
setup-go has a built-in functionality for caching and restoring go modules and build outputs.
The action defaults to search for the dependency file - go.sum in the repository root, and uses its hash as a part of the cache key.
To use the cache, we should first checkout code. Otherwise action can't find go.mod.

Signed-off-by: Yadong Ding <ding_yadong@foxmail.com>